refactor(memory): split recall-tool into flat helper modules (#350)
Extract response types, Zod schema, filter/trigger helpers, result mapping,
and validation utilities from recall-tool.ts. RecallTool keeps orchestration,
search pipeline, anchor/neighbors, and procedural enrichment unchanged.

Public exports preserved via recall-tool.ts re-exports for package consumers.
